mod-ur: rename, reorg class files
